[engine] Externalize pinocchio algorithms from EngineMultiRobot.
